Fixed reading for generation method
Made a standard for how the text file is supposed to be named depending of which methods that will be used to generate tile blocks. It is implemented so that the generated tile blocks will follow the standard that is written in the file names of the coordinate text files. I also finished the function to to download missing laz files.