Fixed proportion of combine and variation of bridge block
Made it possible to select ratio between positive and negative, segments and block labels. For improved training. Fixed the function that generates tile blocks with bridges, so there is always an bridge point within a square that represent the 75 percent of the area closest to the center of the tile block.